+++ lib/cocoapods/installer/pre_install_hooks_context.rb
@@ -1,0 +1,35 @@
+module Pod
+  class Installer
+    # Context object designed to be used with the HooksManager which describes
+    # the context of the installer before analysis has been completed.
+    #
+    class PreInstallHooksContext
+      # @return [String] The path to the sandbox root (`Pods` directory).
+      #
+      attr_accessor :sandbox_root
+
+      # @return [Podfile] The Podfile for the project.
+      #
+      attr_accessor :podfile
+
+      # @return [Sandbox] The Sandbox for the project.
+      #
+      attr_accessor :sandbox
+
+      # @return [Lockfile] The Lockfile for the project.
+      #
+      attr_accessor :lockfile
+
+      # @return [PreInstallHooksContext] Convenience class method to generate the
+      #         static context.
+      #
+      def self.generate(sandbox, podfile, lockfile)
+        result = new
+        result.podfile = podfile
+        result.sandbox = sandbox
+        result.lockfile = lockfile
+        result
+      end
+    end
+  end
+end